requirements.txt
aws-cdk-lib==2.47.0
constructs>=10.0.0,<11.0.0
boto3==1.24.96
awswrangler==2.17.0
